+++ js/dati.js
@@ -16,7 +16,7 @@
    A `twitch.domini` la generazione aggiunge sempre localhost e 127.0.0.1;
    il player ci mette da sé location.hostname.
 
-   Generato il 2026-09-21T11:24:41.444Z.
+   Generato il 2026-09-22T06:04:52.625Z.
    ===================================================================== */
 window.DATI = {
   "twitch": {